GIX-2060: Move nns fee stores
